backend/requirements.txt
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
fastapi
uvicorn[standard]
python-multipart
pytesseract
torch
torchvision
tensorflow
streamlit
